src/retail/pyproject.toml
[tool.poetry]
name = "retail"
version = "0.1.0"
description = "Synvya Retail Commerce application supporting plugins for Square and Shopify."
authors = ["Alejandro alejandro@synvya.com"]
readme = "README.md"
license = "MIT"
packages = [{include = "retail", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.28.1"}
sqlalchemy = "^2.0.29"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
squareup = "^20.1.0"
synvya_sdk = "^0.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
